services/asr-tts-service/requirements.txt
# ASR-TTS Service 依赖

# FastAPI 相关
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ASR (Whisper)
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.24.0

# TTS
edge-tts>=6.1.0
pyttsx3>=2.90

# 工具库
zhconv>=1.4.2

# 日志
loguru>=0.7.0

